+++ src/Saphir/Declaration/Types.hpp
@@ -232,10 +232,10 @@ namespace EmEn::Saphir::Declaration
 	/**
 	 * @brief Returns the size in bytes from a GLSL type.
 	 * @param type The variable type.
-	 * @return size_t
+	 * @return uint32_t
 	 */
 	[[nodiscard]]
-	size_t size_bytes (VariableType type) noexcept;
+	uint32_t size_bytes (VariableType type) noexcept;
 
 	/**
 	 * @brief The GLSL memory layout enumeration.
